Fix 3 commander bugs: endless mode, boss phase 1 leash, boss player commander
- Export spawnCommander from helpers.ts so both engine.ts and endlessMode.ts can use it
- Endless mode: spawn commanders at game start (was guarded by !endlessMode)
- Endless mode: respawn opponent commander each wave in triggerNextEndlessWave
- Endless mode: spare player commander from wave finger-smash
- Boss phase 1: set isCommander + commanderHomeX on boss unit so it stays leashed
- Boss phase 1: spawn player commander alongside boss unit
- Boss phase 2: clear isCommander/commanderHomeX so boss charges freely
- Boss phase 2 shockwave: exclude player commander from kill pool
